"Let's get started" page is in main.dart.
Can swap to the onboarding view with elevated button class. Still has the fix the
width of the button. Next step is to add the images for the two onboarding views which will use a
page controller to swipe right. Then add the login page. The exercise/training page should be
connected to bluetooth, which Janroy is currently working on.